requirements.txt
Flask
psycopg2
python-dotenv
flask-cors